+++ deepinv/physics/__init__.py
@@ -59,6 +59,10 @@ from .functional.tomography_subsets import (
     split_measurements,
     split_physics,
 )
+from .ultrasound import (
+    UltrafastUltrasound,
+    UltrasoundPlaneWave,
+)
 from .lidar import SinglePhotonLidar
 from .singlepixel import SinglePixelCamera
 from .remote_sensing import Pansharpen